Former-commit-id: 9499bede4a0720956486156cdb493881503b59d9

+++ src/com/vimukti/accounter/core/Utility.java
@@ -181,7 +181,7 @@ public class Utility {
 		return transactionName;
 	}
 
-	public static String getAccountTypeString(int accountingType,
+	public static String getAccountTypeString(long accountingType,
 			int accountType) {
 
 		String accountTypeName = null;
